+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [12.0.0] - 2021-10-16
+
 ### Changed
 
 - Enabled vue/array-bracket-newline rule with `'consistent'`